Staging

+++ Regions/UK_entities/Regional_Run_Files/AWS_calls.py
@@ -6,11 +6,12 @@ import boto3
 from botocore.exceptions import ClientError
 import pdb
 from zipfile import ZipFile
+import pandas as pd
 
 # get the remote database details from .env
-load_dotenv(find_dotenv())
-aws_access_key_id = os.environ.get("aws_access_key_id")
-aws_secret_access_key = os.environ.get("aws_secret_access_key")
+# load_dotenv(find_dotenv())
+# aws_access_key_id = os.environ.get("aws_access_key_id")
+# aws_secret_access_key = os.environ.get("aws_secret_access_key")
 
 
 class AwsTransfers(Main):
@@ -50,19 +51,29 @@ class AwsTransfers(Main):
             excluded_matches_fp = self.directories['excluded_matches'].format(self.region_dir, self.proc_type) + '_' + \
                                   str(self.best_config) + '.csv'
 
+            blacklisted_strings_fp = self.directories['blacklisted_string_matches'].format(self.region_dir)
+
+            stats_file_fp = self.directories['script_performance_stats_file'].format(self.region_dir, self.proc_type)
+
             # Assign zip file which will contain above files
             files_zip = self.unverified_file[:10] + "_files.zip"
 
             with ZipFile(files_zip, 'w') as myzip:
-                myzip.write(stats_fp)
-                myzip.write(filtered_matches_fp)
-                myzip.write(excluded_matches_fp)
+                myzip.write(stats_fp, os.path.basename(stats_fp))
+                myzip.write(filtered_matches_fp,os.path.basename(filtered_matches_fp))
+                myzip.write(excluded_matches_fp, os.path.basename(excluded_matches_fp))
+                myzip.write(blacklisted_strings_fp, os.path.basename(blacklisted_strings_fp))
+                myzip.write(stats_file_fp, os.path.basename(stats_file_fp))
 
             self.upload_file(files_zip, self.bucket, 'UK_entities/Archive/' + files_zip)
 
         # Download verified matches from s3 bucket if prodn_verified argument (production only)
         if self.in_args.prodn_verified:
             self.process_verified_files()
+
+        # Add confirmed matches/non-matches to training file
+        if self.in_args.convert_training:
+            self.runfile_mods.convert_training.ConvertToTraining.convert(self)
 
     def process_verified_files(self):
         """
@@ -72,7 +83,8 @@ class AwsTransfers(Main):
         """
 
         # Scan s3 verified folder for files
-        s3 = boto3.client('s3', aws_access_key_id=aws_access_key_id, aws_secret_access_key=aws_secret_access_key)
+
+        s3 = boto3.client('s3', aws_access_key_id=self.aws_access_key_id, aws_secret_access_key=self.aws_secret_access_key)
         response = s3.list_objects(Bucket=self.bucket, Prefix='UK_entities/Verified_Matches/')
 
         # Ignore first file entry in dict as is just the folder name. Returns a list of files
@@ -80,43 +92,84 @@ class AwsTransfers(Main):
 
         # # For any files in /s3/verified/ - download them to local /verified_matches/
         for i in range(len(files)):
-
+            verified_fp = os.path.join(self.directories['verified_matches_dir'].format(self.region_dir,self.proc_type),os.path.basename(files[i]['Key']))
             s3.download_file(self.bucket,
                              files[i]['Key'],
-                             os.path.join(self.directories['verified_matches_dir'].format(self.region_dir,self.proc_type),os.path.basename(files[i]['Key'])))
+                             verified_fp)
 
-        # # Upload all files in verified_matches_dir to our database:
-        self.runfile_mods.db_calls.DbCalls(self).addDataToTable()
+        # Upload all files in verified_matches_dir to our database:
+        if self.in_args.upload:
+            self.runfile_mods.db_calls.DbCalls(self).addDataToTable()
+
+        # Loop through retrieved verified matches files from S3 bucket
 
         for i in range(len(files)):
             try:
                 # Delete from unverified folder (if hasn't been done by team already) so team know which haven't been
                 # verified yet (located via date prefix of verified file incase of name change by team)
-                response = s3.list_objects(Bucket=self.bucket, Prefix='UK_entities/Unverified_Matches/' + os.path.basename(files[i]['Key'])[:10])
-                file = response['Contents'][:]
-                s3.delete_object(Bucket=self.bucket, Key=file[i]['Key'])
+                if self.in_args.upload:
+
+                    s3.delete_object(Bucket=self.bucket, Key=os.path.join('UK_entities','Unverified_Matches', os.path.basename(files[i]['Key'])))
             except:
                 pass
 
-            # Convert file to zip file for archiving
-            zip_fp = os.path.join(self.directories['verified_matches_dir'].format(self.region_dir, self.proc_type)
-                                    , os.path.basename(files[i]['Key'][:-4] + '.zip'))
+            # For each verified file, iterate over S3 zip files and download the corresponding zip.
+            # Need to iterate over both files in /verified and /archive to make sure we aren't adding the wrong information
+            # to multiple different files when theres >1 file in these folders.
+            s3 = boto3.client('s3', aws_access_key_id=self.aws_access_key_id, aws_secret_access_key=self.aws_secret_access_key)
+            response = s3.list_objects(Bucket=self.bucket, Prefix='UK_entities/Archive/')
+            archive_files = response['Contents'][1:]
 
-            file = os.path.join(self.directories['verified_matches_dir'].format(self.region_dir, self.proc_type)
-                                  , os.path.basename(files[i]['Key']))
+            # For each file in the archive folder, iterate over the names and find any matches to the date of the current
+            # verified file in the iterator
+            verified_zip_name = os.path.basename(files[i]['Key'])[:10] + '_files.zip'
+            for a in range(len(archive_files)):
+                # if they match, open that archive zip file and open the script performance stats file within it
+                # and get the length/count of verified matches from the verified file
+                archive_file = os.path.basename(archive_files[a]['Key'])
+                if archive_file == verified_zip_name:
+                    # Download archive file to local verified folder
+                    dl_archive_fp = os.path.join(os.path.join(self.directories['verified_matches_dir']
+                                                              .format(self.region_dir, self.proc_type), archive_file))
+                    s3.download_file(self.bucket,
+                                     os.path.join('UK_entities/Archive/', archive_file), dl_archive_fp)
 
-            with ZipFile(zip_fp, 'w') as myzip:
-                myzip.write(file)
+                    # Open archive file
+                    with ZipFile(dl_archive_fp, 'r') as z:
 
-            # Upload zip file to S3 Archive
-            self.upload_file(zip_fp, self.bucket, 'UK_entities/Archive/' + os.path.basename(zip_fp))
+                        # Open corresponding verified matches file
+                        verified_fp = os.path.join(
+                            self.directories['verified_matches_dir'].format(self.region_dir, self.proc_type),
+                            os.path.basename(files[i]['Key']))
+                        ver_file = pd.read_csv(verified_fp)
+                        with z.open('script_performance_stats.csv') as f:
+                            # Add additional stats to script performance stats csv
+                            stats_file = pd.read_csv(f)
+                            true_positives = len(ver_file[ver_file['Manual_Match_N'] == 'Y'])
+                            false_positives = len(ver_file[ver_file['Manual_Match_N'] == 'N'])
+                            unverified = len(ver_file) - true_positives - false_positives
+                            stats_file['true_positives'] = true_positives
+                            stats_file['false_positives'] = false_positives
+                            stats_file['script_precision'] = round((true_positives / (false_positives + true_positives)) * 100, 2)
+                            stats_file['unverified'] = unverified
+                            stats_file.to_csv(self.directories['script_performance_stats_file'].format(self.region_dir, self.proc_type)
+                                              ,index=False)
 
-            # Delete matches csv from s3 verified folder
-            s3.delete_object(Bucket=self.bucket, Key=files[i]['Key'])
+                    stats_file_fp = self.directories['script_performance_stats_file'].format(self.region_dir,
+                                                                                             self.proc_type)
+                    with ZipFile(dl_archive_fp, 'a') as z:
+
+                        # Add/overwrite new stats file and verified matches file to zip file, then re-upload to S3 /Archive
+                        z.write(stats_file_fp, os.path.basename(stats_file_fp))
+                        z.write(verified_fp, os.path.basename(verified_fp))
+                        self.upload_file(dl_archive_fp, self.bucket, 'UK_entities/Archive/' + archive_file)
+
+            # Delete matches csv from s3 verified folder (if 'upload' arg used)
+            if self.in_args.upload:
+                s3.delete_object(Bucket=self.bucket, Key=files[i]['Key'])
 
 
-    @staticmethod
-    def upload_file(file_name, bucket, object_name=None):
+    def upload_file(self, file_name, bucket, object_name=None):
         """Upload a file to an S3 bucket
 
         :param file_name: File to upload
@@ -130,7 +183,7 @@ class AwsTransfers(Main):
             object_name = file_name
 
         # Upload the file
-        s3_client = boto3.client('s3', aws_access_key_id=aws_access_key_id, aws_secret_access_key=aws_secret_access_key)
+        s3_client = boto3.client('s3', aws_access_key_id=self.aws_access_key_id, aws_secret_access_key=self.aws_secret_access_key)
         try:
             s3_client.upload_file(file_name, bucket, object_name)
         except ClientError as e:
